backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
langgraph==0.1.5
langchain-anthropic==0.1.13
langchain-core==0.2.5
langchain-community==0.2.4
pdfplumber==0.11.0
httpx==0.27.0
python-dotenv==1.0.1
supabase==2.4.4
pydantic==2.7.1
